must fix profile servlet

+++ LinkedIn/src/Servlets/PostCreation.java
@@ -183,7 +183,7 @@ public class PostCreation extends HttpServlet {
 			
 			//create new post
 			UserDAO userDao = new UserDAOImpl(true);
-			User user = userDao.find(Long.valueOf((String) request.getSession().getAttribute("id")));
+			User user = userDao.find(Integer.valueOf((String) request.getSession().getAttribute("id")));
 			
 			Post newPost = new Post(text,dNow,pathFiles,hasAudio, hasImages, hasVideo, 0, user);
 			dao.create(newPost);
